Move migrate APIs to indices namespace
