+++ specification/_json_spec/indices.cancel_migrate_reindex.json
@@ -1,5 +1,5 @@
 {
-  "migrate.get_reindex_status": {
+  "indices.cancel_migrate_reindex": {
     "documentation": {
       "url": "https://www.elastic.co/guide/en/elasticsearch/reference/master/data-stream-reindex.html",
       "description": "This API returns the status of a migration reindex attempt for a data stream or index"
@@ -13,8 +13,8 @@
     "url": {
       "paths": [
         {
-          "path": "/_migration/reindex/{index}/_status",
-          "methods": ["GET"],
+          "path": "/_migration/reindex/{index}/_cancel",
+          "methods": ["POST"],
           "parts": {
             "index": {
               "type": "string",
