+++ specification/_json_spec/indices.create_from.json
@@ -1,5 +1,5 @@
 {
-  "migrate.create_from": {
+  "indices.create_from": {
     "documentation": {
       "url": "https://www.elastic.co/guide/en/elasticsearch/reference/master/data-stream-reindex.html",
       "description": "This API creates a destination from a source index. It copies the mappings and settings from the source index while allowing request settings and mappings to override the source values."
