+++ specification/_json_spec/indices.get_migrate_reindex_status.json
@@ -1,5 +1,5 @@
 {
-  "migrate.cancel_reindex": {
+  "indices.get_migrate_reindex_status": {
     "documentation": {
       "url": "https://www.elastic.co/guide/en/elasticsearch/reference/master/data-stream-reindex.html",
       "description": "This API returns the status of a migration reindex attempt for a data stream or index"
@@ -13,8 +13,8 @@
     "url": {
       "paths": [
         {
-          "path": "/_migration/reindex/{index}/_cancel",
-          "methods": ["POST"],
+          "path": "/_migration/reindex/{index}/_status",
+          "methods": ["GET"],
           "parts": {
             "index": {
               "type": "string",
