+++ specification/_json_spec/indices.migrate_reindex.json
@@ -1,5 +1,5 @@
 {
-  "migrate.reindex": {
+  "indices.migrate_reindex": {
     "documentation": {
       "url": "https://www.elastic.co/guide/en/elasticsearch/reference/master/data-stream-reindex.html",
       "description": "This API reindexes all legacy backing indices for a data stream. It does this in a persistent task. The persistent task id is returned immediately, and the reindexing work is completed in that task"
