+++ specification/indices/cancel_migrate_reindex/MigrateCancelReindexRequest.ts
@@ -18,36 +18,20 @@
  */
 
 import { RequestBase } from '@_types/Base'
-import { IndexName } from '@_types/common'
+import { Indices } from '@_types/common'
 
 /**
- * "This API reindexes all legacy backing indices for a data stream. It does this in a persistent task. The persistent task id is returned immediately, and the reindexing work is completed in that task
+ * This API cancels a migration reindex attempt for a data stream or index
  *
- * @rest_spec_name migrate.reindex
+ * @rest_spec_name indices.cancel_migrate_reindex
  * @availability stack since=8.18.0 stability=experimental
+ * @availability serverless stability=experimental visibility=private
  * @doc_id migrate
  * @doc_tag migration
  */
 export interface Request extends RequestBase {
-  /** @codegen_name reindex */
-  body: MigrateReindex
-}
-
-export class MigrateReindex {
-  /**
-   * Reindex mode. Currently only 'upgrade' is supported.
-   */
-  mode: ModeEnum
-  /**
-   * The source index or data stream (only data streams are currently supported).
-   */
-  source: SourceIndex
-}
-
-export class SourceIndex {
-  index: IndexName
-}
-
-enum ModeEnum {
-  upgrade
+  path_parts: {
+    /** The index or data stream name */
+    index: Indices
+  }
 }
