+++ specification/indices/create_from/MigrateCreateFromRequest.ts
@@ -25,7 +25,7 @@ import { TypeMapping } from '@_types/mapping/TypeMapping'
 /**
  * This API creates a destination from a source index. It copies the mappings and settings from the source index while allowing request settings and mappings to override the source values.
  *
- * @rest_spec_name migrate.create_from
+ * @rest_spec_name indices.create_from
  * @availability stack since=8.18.0 stability=experimental
  * @availability serverless stability=experimental visibility=private
  * @doc_id migrate
