+++ specification/indices/migrate_reindex/MigrateReindexRequest.ts
@@ -18,20 +18,36 @@
  */
 
 import { RequestBase } from '@_types/Base'
-import { Indices } from '@_types/common'
+import { IndexName } from '@_types/common'
 
 /**
- * This API returns the status of a migration reindex attempt for a data stream or index
+ * "This API reindexes all legacy backing indices for a data stream. It does this in a persistent task. The persistent task id is returned immediately, and the reindexing work is completed in that task
  *
- * @rest_spec_name migrate.get_reindex_status
+ * @rest_spec_name indices.migrate_reindex
  * @availability stack since=8.18.0 stability=experimental
- * @availability serverless stability=experimental visibility=private
  * @doc_id migrate
  * @doc_tag migration
  */
 export interface Request extends RequestBase {
-  path_parts: {
-    /** The index or data stream name */
-    index: Indices
-  }
+  /** @codegen_name reindex */
+  body: MigrateReindex
+}
+
+export class MigrateReindex {
+  /**
+   * Reindex mode. Currently only 'upgrade' is supported.
+   */
+  mode: ModeEnum
+  /**
+   * The source index or data stream (only data streams are currently supported).
+   */
+  source: SourceIndex
+}
+
+export class SourceIndex {
+  index: IndexName
+}
+
+enum ModeEnum {
+  upgrade
 }
